Removed windows tests. Restructure validations.
So there is no `ob validate` but, rather, each subcommand will have `validate`s, `check`s, etc. Added the 'validate yaml' to the `run` typer.
